ReadinessCheck - this will not work
Maybe if we had applied a `Job` to the cluster directly, this would have
worked, but the crossplane resource was a Kubernetes Provider Object

We can't expect Object to mirror the status of the Job that it created
We could program it with a CEL Query expression though, I think?

I'm not going to do that.

Signed-off-by: Kingdon B <kingdon@urmanac.com>