first stab at "myapp" composition use
we will have to work on a way that secrets do not get embedded in the
body of the composition claim spec

This exploration also showed that there are some limitations to the way
that strings can be substituted. For example, envValue is easy, envKey
is not so easy. Maybe we could patch a map[string]string named env into
the deployment env? (or into a generated secret, as we're doing here)

Signed-off-by: Kingdon B <kingdon@urmanac.com>